% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{midpoints}
\alias{midpoints}
\title{Midpoints of a Numeric Vector}
\usage{
midpoints(x)
}
\arguments{
\item{x}{A numeric vector with at least 2 elements.}
}
\value{
A numeric vector of \link{length} \code{length(x) - 1}.
}
\description{
Returns the midpoints between the elements of a vector in the order the
elements appear.
}
\details{
This function does not sort.
}
\examples{
midpoints(c(0, 1, 10, 4))
}
